refactor: move pena link consumption slice into core application
