+++ backend/src/persistence/application/use_cases/__init__.py
@@ -22,6 +22,12 @@ from core.application.use_cases.get_player_profile_usecase import GetPlayerProfi
 from core.application.use_cases.get_season_match_insights_usecase import (
     GetSeasonMatchInsightsUseCase,
 )
+from core.application.use_cases.link_user_to_pena_usecase import (
+    InvalidLinkTokenError,
+    LinkUserToPenaUseCase,
+    UserAlreadyLinkedError,
+    UserProfileNotFoundError,
+)
 from core.application.use_cases.manage_pena_labels_usecase import (
     InvalidPenaLabelsDataError,
     ManagePenaLabelsUseCase,
@@ -47,12 +53,6 @@ from core.application.use_cases.update_player_profile_usecase import (
 )
 
 from .get_player_profile_usecase import PenaInfo
-from .link_user_to_pena_usecase import (
-    InvalidLinkTokenError,
-    LinkUserToPenaUseCase,
-    UserAlreadyLinkedError,
-    UserProfileNotFoundError,
-)
 from .manage_pena_accountability_usecase import (
     InvalidPenaAccountabilityDataError,
     ManagePenaAccountabilityUseCase,
